tegra: Make BO ref / unref return the refcount value on success
For a debugging purposes, it should be useful to know the current BO
reference count number. Let's make ref / unref return that count on success,
note that since unref's count is potentially unreliable due to a lack of
locking around atomic_read(), but realistically that race condition is
nearly impossible, it is good enough for debugging purposes.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>