tegra: Avoid BO counter wraparound
We are resetting mmap refcounter for the taken out from cache BO, but
BO unmap function checks for the map pointer rather than for mmap count.
The mmap_ref==0 + map!=NULL is the case for a BO taken from cache, mmap_ref
will wraparound if BO is unmapped without mapping it at first.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>