Remove all libdrm packages except libdrm-tegra
Recent Opentegra, libvdpau-tegra, mesa now do not depend on libdrm-tegra,
but older mesa (which is used by older distro versions) uses obsolete grate's
libdrm-tegra. Mesa will continue to work on older distros, but it will be
better for everyone to upgrade to a recent distro version for the best
experience.

Now only libdrm-tegra package is provided by the grate-driver PPA for older
distro versions. It doesn't include sources for libdrm-tegra, upgrade to
Ubuntu 20.04 or use git sources by yourself if you need it.